fix: nDPI 4.12 protocol detection not working - missing protocol bitmask setup
The nDPI 4.12 API requires calling ndpi_set_protocol_detection_bitmask2() to enable
protocol detection after initializing the detection module. Without this, all protocols
remain disabled and nDPI returns UNKNOWN for all traffic.

Changes:
- classifier/ndpi_linux.c: Add ndpi_set_protocol_detection_bitmask2() call
- proxy/ndpi_linux.c: Update to use nDPI 4.12 API (ndpi_proto.proto.*, client_packet_direction)
- proxy/ndpi_linux.h: Fix category type from uint8 to uint32
- proxy/Dockerfile: Build nDPI 4.12 from source instead of using system lib
- proxy/ndpi.go: Update CGO paths to use /usr/local/include

+++ proxy/internal/ndpi/gondpi/ndpi_linux.h
@@ -4,7 +4,7 @@
 typedef struct ndpi_proto_result {
     uint16_t master_protocol;
     uint16_t app_protocol;
-    uint8_t category;
+    uint32_t category;
 } ndpi_proto_result_t;
 
 extern ndpi_proto_result_t ndpi_detection_process_wrapper(struct ndpi_detection_module_struct *,
